+++ terrain/variable_lod/voxel_mesh_block_vlt.cpp
@@ -45,7 +45,8 @@ VoxelMeshBlockVLT::~VoxelMeshBlockVLT() {
 	}
 }
 
-void VoxelMeshBlockVLT::set_mesh(Ref<Mesh> mesh, DirectMeshInstance::GIMode gi_mode) {
+void VoxelMeshBlockVLT::set_mesh(
+		Ref<Mesh> mesh, DirectMeshInstance::GIMode gi_mode, RenderingServer::ShadowCastingSetting shadow_casting) {
 	// TODO Don't add mesh instance to the world if it's not visible.
 	// I suspect Godot is trying to include invisible mesh instances into the culling process,
 	// which is killing performance when LOD is used (i.e many meshes are in pool but hidden)
@@ -56,6 +57,7 @@ void VoxelMeshBlockVLT::set_mesh(Ref<Mesh> mesh, DirectMeshInstance::GIMode gi_m
 			// Create instance if it doesn't exist
 			_mesh_instance.create();
 			_mesh_instance.set_gi_mode(gi_mode);
+			_mesh_instance.set_cast_shadows_setting(shadow_casting);
 			set_mesh_instance_visible(_mesh_instance, _visible && _parent_visible);
 		}
 
@@ -86,7 +88,18 @@ void VoxelMeshBlockVLT::set_gi_mode(DirectMeshInstance::GIMode mode) {
 	}
 }
 
-void VoxelMeshBlockVLT::set_transition_mesh(Ref<Mesh> mesh, unsigned int side, DirectMeshInstance::GIMode gi_mode) {
+void VoxelMeshBlockVLT::set_shadow_casting(RenderingServer::ShadowCastingSetting mode) {
+	VoxelMeshBlock::set_shadow_casting(mode);
+	for (unsigned int i = 0; i < _transition_mesh_instances.size(); ++i) {
+		DirectMeshInstance &mi = _transition_mesh_instances[i];
+		if (mi.is_valid()) {
+			mi.set_cast_shadows_setting(mode);
+		}
+	}
+}
+
+void VoxelMeshBlockVLT::set_transition_mesh(Ref<Mesh> mesh, unsigned int side, DirectMeshInstance::GIMode gi_mode,
+		RenderingServer::ShadowCastingSetting shadow_casting) {
 	DirectMeshInstance &mesh_instance = _transition_mesh_instances[side];
 
 	if (mesh.is_valid()) {
@@ -94,6 +107,7 @@ void VoxelMeshBlockVLT::set_transition_mesh(Ref<Mesh> mesh, unsigned int side, D
 			// Create instance if it doesn't exist
 			mesh_instance.create();
 			mesh_instance.set_gi_mode(gi_mode);
+			mesh_instance.set_cast_shadows_setting(shadow_casting);
 			set_mesh_instance_visible(mesh_instance, _visible && _parent_visible && _is_transition_visible(side));
 		}