+++ terrain/voxel_node.cpp
@@ -173,6 +173,17 @@ VoxelNode::GIMode VoxelNode::get_gi_mode() const {
 	return _gi_mode;
 }
 
+void VoxelNode::set_shadow_casting(GeometryInstance3D::ShadowCastingSetting setting) {
+	if (setting != _shadow_casting) {
+		_shadow_casting = setting;
+		_on_shadow_casting_changed();
+	}
+}
+
+GeometryInstance3D::ShadowCastingSetting VoxelNode::get_shadow_casting() const {
+	return _shadow_casting;
+}
+
 void VoxelNode::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("set_stream", "stream"), &VoxelNode::_b_set_stream);
 	ClassDB::bind_method(D_METHOD("get_stream"), &VoxelNode::_b_get_stream);
@@ -186,6 +197,9 @@ void VoxelNode::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("set_gi_mode", "mode"), &VoxelNode::set_gi_mode);
 	ClassDB::bind_method(D_METHOD("get_gi_mode"), &VoxelNode::get_gi_mode);
 
+	ClassDB::bind_method(D_METHOD("set_shadow_casting", "mode"), &VoxelNode::set_shadow_casting);
+	ClassDB::bind_method(D_METHOD("get_shadow_casting"), &VoxelNode::get_shadow_casting);
+
 	ADD_PROPERTY(PropertyInfo(Variant::OBJECT, "stream", PROPERTY_HINT_RESOURCE_TYPE, VoxelStream::get_class_static()),
 			"set_stream", "get_stream");
 	ADD_PROPERTY(
@@ -195,6 +209,8 @@ void VoxelNode::_bind_methods() {
 			"set_mesher", "get_mesher");
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "gi_mode", PROPERTY_HINT_ENUM, "Disabled,Baked,Dynamic"), "set_gi_mode",
 			"get_gi_mode");
+	ADD_PROPERTY(PropertyInfo(Variant::INT, "cast_shadow", PROPERTY_HINT_ENUM, "Off,On,Double-Sided,Shadows Only"),
+			"set_cast_shadows", "get_cast_shadows");
 
 	BIND_ENUM_CONSTANT(GI_MODE_DISABLED);
 	BIND_ENUM_CONSTANT(GI_MODE_BAKED);