+++ util/godot/classes/geometry_instance_3d.h
@@ -1,0 +1,11 @@
+#ifndef ZN_GODOT_GEOMETRY_INSTANCE_3D_H
+#define ZN_GODOT_GEOMETRY_INSTANCE_3D_H
+
+#if defined(ZN_GODOT)
+#include <scene/3d/visual_instance_3d.h>
+#elif defined(ZN_GODOT_EXTENSION)
+#include <godot_cpp/classes/geometry_instance3d.hpp>
+using namespace godot;
+#endif
+
+#endif // ZN_GODOT_GEOMETRY_INSTANCE_3D_H